packages/utils/manifest_get/requirements.txt
PyYAML==5.3
